requirements.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
binaryornot==0.4.4
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==45.0.6
cssselect2==0.8.0
cycler==0.12.1
Django==5.2.1
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-tailwind==4.2.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.59.0
html5lib==1.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.0
mariadb==1.1.13
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mdurl==0.1.2
mysqlclient==2.2.7
numpy==2.3.5
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyparsing==3.2.5
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20250708
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
zopfli==0.2.3.post1
